requirements.txt
langchain>=0.2.12
langchain-core>=0.2.30
langchain-community>=0.2.14
langchain-huggingface>=0.1.2
langchain-chroma>=0.1.5
langchain-deepseek>=0.1.1
chromadb>=0.5.5
sentence-transformers>=3.0.1
torch>=2.0.0,<3.0.0
tqdm>=4.66.0
unstructured>=0.14.0
pypdf>=4.0.0
gradio>=4.18.0
python-dotenv>=1.0.0